server/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
audioread==3.0.1
autocorrect==2.6.1
av==14.3.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.2
capstone==6.0.0a3
catboost==1.2.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colored-traceback==0.4.2
colorlog==6.9.0
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.0
cryptography==44.0.0
cvzone==1.6.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
Django==5.1.5
djangorestframework==3.15.2
dnspython==2.7.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.19.1
easyocr==1.7.2
email_validator==2.2.0
enchant==0.0.1
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
fonttools==4.54.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
gotrue==2.12.0
gradio==5.12.0
gradio_client==1.5.4
graphviz==0.20.3
greenlet==3.1.1
groq==0.15.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hashcat==24.11.10
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
imbalanced-learn==0.13.0
imblearn==0.0
imutils==0.5.4
inexactsearch==1.0.2
iniconfig==2.1.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.2
kagglehub==0.3.4
keras==3.9.2
kiwisolver==1.4.7
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.31
langchain-groq==0.2.3
langchain-text-splitters==0.3.5
langsmith==0.2.11
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightgbm==4.5.0
lightning-utilities==0.14.3
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.0.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
namex==0.0.8
narwhals==1.35.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
noisereduce==3.0.3
notebook_shim==0.2.4
numba==0.60.0
numpy==2.2.4
ollama==0.4.7
openai==1.72.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
optuna==4.2.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
patsy==1.0.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
plumbum==1.9.0
pooch==1.8.2
postgrest==1.0.1
proglog==0.1.11
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
pwn==1.0
pwntools==4.14.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pyelftools==0.31
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
PyNaCl==1.5.0
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
pyproj==3.7.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.5
pytest-mock==3.14.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pyttsx3==2.98
pytube==15.0.0
pytweening==1.2.0
pytz==2025.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rich-toolkit==0.13.2
ROPGadget==7.6
rpds-py==0.24.0
rpyc==6.0.1
rsa==4.9
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==78.1.0
shapely==2.1.0
shellingham==1.5.4
silpa_common==0.3
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundex==1.1.3
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spellchecker==0.4
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
storage3==0.11.3
streamlit==1.44.1
streamlit-option-menu==0.4.0
streamlit-webrtc==0.62.4
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.18.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.5.1+cu118
torchmetrics==1.7.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
unicorn==2.1.1
unix-ar==0.2.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
virtualenv==20.27.1
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.3
wrapt==1.17.2
WTForms==3.2.1
xgboost==2.1.2
yarl==1.19.0
yt-dlp==2025.4.30
zstandard==0.23.0
